fix: consolidate runtimetests to satisfy pkg-file-count limit
Merge fresh-init and legacy-encoded untouched fixtures into named_factory_resolution_test.go and named_factory_list_test.go so pkg/config/runtimetests stays at 15 files.
